requirements.txt
chainlit
langchain
openai
pathlib
gitpython
langchain_openai
langchain_community
httpx
